Add a getter for the inner tracing-subscriber's Layer
Before this commit, the only way to effectively consume the layer
created with this crate is to use the `subscriber()` method that wraps
the layer into a `Registry`. This might not always be desired by end
users of this crate, so this commit adds a new way to retrieve the inner
tracing's `Layer` so one can add it into an existing Registry, for
instance.